moved helper types
