+++ packages/sit-onyx/src/components/OnyxDataGrid/types.ts
@@ -1,27 +1,7 @@
+import type { KeysOfUnion, MaybePick, UnionByKey } from "../../types";
 import type { ColumnConfig, DataGridFeature, TypeRenderMap } from "./features";
 
 export type DataGridMetadata = Record<string, unknown>;
-
-/**
- * Gets all possible keys in a union.
- */
-export type KeysOfUnion<T> = T extends T ? keyof T : never;
-
-/**
- * Merge the values for all entries by key.
- */
-export type UnionByKey<T> = {
-  [Key in T extends NonNullable<T> ? keyof T : never]: T extends { [P in Key]?: infer Value }
-    ? Value extends undefined
-      ? never
-      : Value
-    : never;
-};
-
-/**
- * Pick a value form `T` for the key `K`, if it exists.
- */
-export type MaybePick<T, K> = K extends keyof T ? T[K] : never;
 
 /**
  * Unwraps the defined typeRenderers
